+++ Tests/MapboxMapsTests/Annotations/ViewAnnotationManagerTests.swift
@@ -11,7 +11,9 @@ final class ViewAnnotationManagerTests: XCTestCase {
         super.setUp()
         container = UIView()
         mapboxMap = MockMapboxMap()
-        manager = ViewAnnotationManager(containerView: container, mapboxMap: mapboxMap)
+        manager = ViewAnnotationManager(
+            containerView: container,
+            mapboxMap: mapboxMap)
     }
 
     override func tearDown() {
@@ -410,6 +412,35 @@ final class ViewAnnotationManagerTests: XCTestCase {
 
         XCTAssertTrue(observer.framesDidChangeStub.invocations.isEmpty)
         XCTAssertTrue(observer.visibilityDidChangeStub.invocations.isEmpty)
+    }
+
+    func testCameraForAnnotations() throws {
+        let points: [CLLocationCoordinate2D] = .random(withLength: 4, generator: CLLocationCoordinate2D.random)
+        for (index, point) in points.enumerated() {
+            let options = ViewAnnotationOptions(geometry: Point(point).geometry, width: 40, height: 40)
+            try manager.add(UIView(), id: "\(index)", options: options)
+            mapboxMap.optionsForViewAnnotationWithIdStub.returnValueQueue.insert(options, at: 0)
+        }
+
+        let padding = UIEdgeInsets.random()
+        let bearing = CGFloat.random(in: -180...180)
+        let pitch = CGFloat.random(in: 0...90)
+        _ = manager.camera(forAnnotations: ["0", "1", "2", "3"], padding: padding, bearing: bearing, pitch: pitch)
+
+        let parameters = try XCTUnwrap(mapboxMap.cameraForGeometryStub.invocations.last).parameters
+        XCTAssertEqual(parameters.bearing, bearing)
+        XCTAssertEqual(parameters.pitch, pitch)
+
+        let coordinates = try XCTUnwrap(MapboxCommon.Geometry(parameters.geometry).extractLocationsArray()).map(\.mkCoordinateValue)
+        let north = try XCTUnwrap(coordinates.max(by: { $0.latitude < $1.latitude })).latitude
+        let east = try XCTUnwrap(coordinates.max(by: { $0.longitude < $1.longitude })).longitude
+        let south = try XCTUnwrap(coordinates.min(by: { $0.latitude < $1.latitude })).latitude
+        let west = try XCTUnwrap(coordinates.min(by: { $0.longitude < $1.longitude })).longitude
+
+        XCTAssertFalse(points.contains(where: { $0.latitude > north }))
+        XCTAssertFalse(points.contains(where: { $0.longitude > east }))
+        XCTAssertFalse(points.contains(where: { $0.latitude < south }))
+        XCTAssertFalse(points.contains(where: { $0.longitude < west }))
     }
 
     // MARK: - Helper functions